Move context menu away from preload
